[code] move util into src
