+++ packages/argdown-vscode/src/util/IDictionary.ts
