+++ packages/argdown-vscode/src/util/arrays.ts
