+++ packages/argdown-vscode/src/util/dispose.ts
