+++ packages/argdown-vscode/src/util/file.ts
