+++ packages/argdown-vscode/src/util/path.ts
