+++ packages/argdown-vscode/src/util/resources.ts
