+++ packages/argdown-vscode/src/util/topmostLineMonitor.ts
